- usunięcie niektórych warningów
git-svn-id: svn://svn.nlp.ipipan.waw.pl/morfeusz/morfeusz@26 ff4e3ee1-f430-4e82-ade0-24591c43f1fd

+++ morfeusz/FlexionGraph.hpp
@@ -18,7 +18,7 @@ public:
 
     struct Edge {
         InterpretedChunk chunk;
-        int nextNode;
+        unsigned long nextNode;
     };
 
     void addPath(const std::vector<InterpretedChunk>& path);
